+++ src/commonMain/kotlin/dev/inmo/krontab/Executes.kt
@@ -6,6 +6,7 @@ import kotlinx.coroutines.delay
 import kotlinx.coroutines.isActive
 import kotlin.coroutines.coroutineContext
 
+
 /**
  * Execute [block] once at the [KronScheduler.next] time and return result of [block] calculation.
  *
@@ -13,12 +14,22 @@ import kotlin.coroutines.coroutineContext
  *
  * WARNING!!! In case if [KronScheduler.next] of [this] instance will return null, [block] will be called immediately
  */
-suspend inline fun <T> KronScheduler.doOnceLocal(noinline block: suspend (DateTime) -> T): T {
+suspend inline fun <T> KronScheduler.doOnce(noinline block: suspend (DateTime) -> T): T {
     val time = nextOrNow().also {
         delay((it - DateTime.now()).millisecondsLong)
     }
     return block(time)
 }
+
+/**
+ * Execute [block] once at the [KronScheduler.next] time and return result of [block] calculation.
+ *
+ * WARNING!!! If you want to launch it in parallel, you must do this explicitly.
+ *
+ * WARNING!!! In case if [KronScheduler.next] of [this] instance will return null, [block] will be called immediately
+ */
+@Deprecated("Replaceable", ReplaceWith("doOnce", "dev.inmo.krontab.doOnce"))
+suspend inline fun <T> KronScheduler.doOnceLocal(noinline block: suspend (DateTime) -> T): T = doOnce(block)
 
 /**
  * Execute [block] once at the [KronScheduler.next] time and return result of [block] calculation.
@@ -37,25 +48,16 @@ suspend inline fun <T> KronScheduler.doOnceTz(noinline block: suspend (DateTimeT
 }
 
 /**
- * Execute [block] once at the [KronScheduler.next] time and return result of [block] calculation.
- *
- * WARNING!!! If you want to launch it in parallel, you must do this explicitly.
- *
- * WARNING!!! In case if [KronScheduler.next] of [this] instance will return null, [block] will be called immediately
- */
-suspend inline fun <T> KronScheduler.doOnce(noinline block: suspend () -> T): T = doOnceLocal { _ -> block() }
-
-/**
- * Will [buildSchedule] using [scheduleConfig] and call [doOnceLocal] on it
+ * Will [buildSchedule] using [scheduleConfig] and call [doOnce] on it
  * @see buildSchedule
  */
 suspend inline fun <T> doOnce(
     scheduleConfig: String,
     noinline block: suspend (DateTime) -> T
-) = buildSchedule(scheduleConfig).doOnceLocal(block)
+) = buildSchedule(scheduleConfig).doOnce(block)
 
 /**
- * Will [buildSchedule] using [scheduleConfig] and call [doOnceLocal] on it
+ * Will [buildSchedule] using [scheduleConfig] and call [doOnce] on it
  * @see buildSchedule
  */
 suspend inline fun <T> doOnceTz(
@@ -63,24 +65,20 @@ suspend inline fun <T> doOnceTz(
     noinline block: suspend (DateTimeTz) -> T
 ) = buildSchedule(scheduleConfig).doOnceTz(block)
 
-/**
- * Will [buildSchedule] using [scheduleConfig] and call [doOnceLocal] on it
- * @see buildSchedule
- */
-suspend inline fun <T> doOnce(
-    scheduleConfig: String,
-    noinline block: suspend () -> T
-) = doOnce(scheduleConfig) { _ -> block() }
-
 
 /**
  * Will execute [block] while it will return true as a result of its calculation
  */
-suspend inline fun KronScheduler.doWhileLocal(noinline block: suspend (DateTime) -> Boolean) {
+suspend inline fun KronScheduler.doWhile(noinline block: suspend (DateTime) -> Boolean) {
     do {
         delay(1L)
-    } while (doOnceLocal(block))
+    } while (doOnce(block))
 }
+/**
+ * Will execute [block] while it will return true as a result of its calculation
+ */
+@Deprecated("Replaceable", ReplaceWith("doWhile", "dev.inmo.krontab.doWhile"))
+suspend inline fun KronScheduler.doWhileLocal(noinline block: suspend (DateTime) -> Boolean) = doWhile(block)
 
 /**
  * Will execute [block] while it will return true as a result of its calculation
@@ -92,19 +90,25 @@ suspend inline fun KronScheduler.doWhileTz(noinline block: suspend (DateTimeTz) 
 }
 
 /**
- * Will execute [block] while it will return true as a result of its calculation
+ * Will [buildSchedule] using [scheduleConfig] and call [doWhile] with [block]
+ *
+ * @see buildSchedule
  */
-suspend inline fun KronScheduler.doWhile(noinline block: suspend () -> Boolean) = doWhileLocal { block() }
+suspend inline fun doWhile(
+    scheduleConfig: String,
+    noinline block: suspend (DateTime) -> Boolean
+) = buildSchedule(scheduleConfig).doWhile(block)
 
 /**
  * Will [buildSchedule] using [scheduleConfig] and call [doWhile] with [block]
  *
  * @see buildSchedule
  */
+@Deprecated("Replaceable", ReplaceWith("doWhile", "dev.inmo.krontab.doWhile"))
 suspend inline fun doWhileLocal(
     scheduleConfig: String,
     noinline block: suspend (DateTime) -> Boolean
-) = buildSchedule(scheduleConfig).doWhileLocal(block)
+) = doWhile(scheduleConfig, block)
 
 /**
  * Will [buildSchedule] using [scheduleConfig] and call [doWhile] with [block]
@@ -116,24 +120,19 @@ suspend inline fun doWhileTz(
     noinline block: suspend (DateTimeTz) -> Boolean
 ) = buildSchedule(scheduleConfig).doWhileTz(block)
 
-/**
- * Will [buildSchedule] using [scheduleConfig] and call [doWhile] with [block]
- *
- * @see buildSchedule
- */
-suspend inline fun doWhile(
-    scheduleConfig: String,
-    noinline block: suspend () -> Boolean
-) = doWhileLocal(scheduleConfig) { block() }
-
 
 /**
  * Will execute [block] without any checking of result
  */
-suspend inline fun KronScheduler.doInfinityLocal(noinline block: suspend (DateTime) -> Unit) = doWhileLocal {
+suspend inline fun KronScheduler.doInfinity(noinline block: suspend (DateTime) -> Unit) = doWhile {
     block(it)
     coroutineContext.isActive
 }
+/**
+ * Will execute [block] without any checking of result
+ */
+@Deprecated("Replaceable", ReplaceWith("doInfinity", "dev.inmo.krontab.doInfinity"))
+suspend inline fun KronScheduler.doInfinityLocal(noinline block: suspend (DateTime) -> Unit) = doInfinity(block)
 
 /**
  * Will execute [block] without any checking of result
@@ -144,22 +143,25 @@ suspend inline fun KronScheduler.doInfinityTz(noinline block: suspend (DateTimeT
 }
 
 /**
- * Will execute [block] without any checking of result
+ * Will [buildSchedule] using [scheduleConfig] and call [doInfinity] with [block]
+ *
+ * @see buildSchedule
  */
-suspend inline fun KronScheduler.doInfinity(noinline block: suspend () -> Unit) = doWhile {
-    block()
-    coroutineContext.isActive
-}
+suspend inline fun doInfinity(
+    scheduleConfig: String,
+    noinline block: suspend (DateTime) -> Unit
+) = buildSchedule(scheduleConfig).doInfinity(block)
 
 /**
  * Will [buildSchedule] using [scheduleConfig] and call [doInfinity] with [block]
  *
  * @see buildSchedule
  */
+@Deprecated("Replaceable", ReplaceWith("doInfinity", "dev.inmo.krontab.doInfinity"))
 suspend inline fun doInfinityLocal(
     scheduleConfig: String,
     noinline block: suspend (DateTime) -> Unit
-) = buildSchedule(scheduleConfig).doInfinityLocal(block)
+) = doInfinity(scheduleConfig, block)
 
 /**
  * Will [buildSchedule] using [scheduleConfig] and call [doInfinity] with [block]
@@ -170,13 +172,3 @@ suspend inline fun doInfinityTz(
     scheduleConfig: String,
     noinline block: suspend (DateTimeTz) -> Unit
 ) = buildSchedule(scheduleConfig).doInfinityTz(block)
-
-/**
- * Will [buildSchedule] using [scheduleConfig] and call [doInfinity] with [block]
- *
- * @see buildSchedule
- */
-suspend inline fun doInfinity(
-    scheduleConfig: String,
-    noinline block: suspend () -> Unit
-) = buildSchedule(scheduleConfig).doInfinity(block)